fix(tokenizer): update tokenizer usage across multiple files
refactor(tokenizer): replace open_clip tokenizer with CLIPTokenizer and add padding/truncation
update tokenizer calls to include padding, truncation and max_length params
modify text encoding to use int32 tensors for ONNX compatibility
add new export configuration with max_text_len=40

+++ pipeline/compile_and_profile.py
@@ -99,7 +99,7 @@ def main():
             onnx_txt_model,
             model_name + f"_text_encoder{postfix}",
             target_device,
-            {"text": ((1, 77), "int64")},
+            {"text": ((1, 77), "int32")},
             "--target_runtime qnn_dlc --truncate_64bit_io",
         )
         img_id = img_compile_future.result()
